requirements.txt
numpy
scipy
pandas
matplotlib
scikit-learn
hyperopt
cma
hpbandster
git+https://github.com/SheffieldML/GPyOpt.git@master
smac[all]
lightgbm
NiaPy==2.0.0rc5
click
google-cloud-storage